nodeadm(feat): add mime multi-part support to file config provider
