--- nodeadm/example/node.eks.aws_v1alpha1_nodeconfig.yaml
@@ -1,23 +1,0 @@
----
-apiVersion: node.eks.aws/v1alpha1
-kind: NodeConfig
-metadata:
-  name: example
-spec:
-  cluster:
-    name: my-cluster
-    apiServerEndpoint: https://example.com
-    certificateAuthority: Y2VydGlmaWNhdGVBdXRob3JpdHk=
-    cidr: 10.100.0.0/16
-  kubelet:
-    config:
-      kind: KubeletConfiguration
-      apiVersion: kubelet.config.k8s.io/v1beta1
-      maxPods: 42
-      clusterDNS:
-        - 0.0.0.0
-        - 1.1.1.1
-    flags:
-      - --v=2
-      - --register-with-taints=foo=bar:NoSchedule
-      - --node-labels=foo=bar
