+++ nodeadm/test/e2e/cases/containerd-runtime-config-neuron/run.sh
@@ -6,7 +6,7 @@ set -o pipefail
 
 source /helpers.sh
 
-mock::aws /etc/aemm-inf1-config.json
+mock::aws aemm-inf1-config.json
 mock::kubelet 1.27.0
 wait::dbus-ready
