+++ nodeadm/test/e2e/cases/containerd-runtime-config-nvidia/run.sh
@@ -6,7 +6,7 @@ set -o pipefail
 
 source /helpers.sh
 
-mock::aws /etc/aemm-g5-config.json
+mock::aws aemm-g5-config.json
 mock::kubelet 1.27.0
 wait::dbus-ready
